+++ src/audio/drc/drc_plat_conf.h
